// desktop/src/operations/mutations/node-collaborator-create.ts
export type NodeCollaboratorCreateMutationInput = {
  type: 'node_collaborator_create';
  userId: string;
  nodeId: string;
  collaboratorIds: string[];
  role: string;
};

export type NodeCollaboratorCreateMutationOutput = {
  success: boolean;
};

declare module '@/operations/mutations' {
  interface MutationMap {
    node_collaborator_create: {
      input: NodeCollaboratorCreateMutationInput;
      output: NodeCollaboratorCreateMutationOutput;
    };
  }
}